Made the recommended changes to getSD including moving it to the util file, dealing with different unit types, ignoring the time axis, adding relevant tests for all of the above, and hopefully cleaning up some of the redundant comments
