+++ test/runtests.jl
@@ -1,10 +1,9 @@
 using ImageMagick
 using RegisterQD
 using Test
-# 
-# include("util.jl")
-# include("qd_random.jl")
-# include("qd_standard.jl")
-# include("gridsearch.jl")
-# include("initial_tfm.jl")
-include("SD_calculator_test.jl")
+
+include("util.jl")
+include("qd_random.jl")
+include("qd_standard.jl")
+include("gridsearch.jl")
+include("initial_tfm.jl")
